refactor: changes files structure
