+++ src/actions/user/index.ts
@@ -2,7 +2,7 @@ import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
 
 import type { iUserState } from './user.types';
-import type { iRootState } from '../redux/stores';
+import type { iRootState } from '../../redux/stores';
 
 const initialState: iUserState = {
   username: null,
