+++ src/components/partials/Header/index.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 
 import { IcArrowRight } from '../../icons';
 import { useAppDispatch } from '../../../redux/hooks';
-import { logout } from '../../../actions/user.slice';
+import { logout } from '../../../actions/user';
 import styles  from './header.module.scss';
 
 export function Header({children}: {children: React.ReactNode}) {
